Cleanup API Routes with Extra Logic
