+++ apps/api/app/api/markets/[id]/balance/route.ts
@@ -1,9 +1,8 @@
 import { NextResponse } from 'next/server'
-import { getAccountBalance } from '@play-money/accounts/lib/getAccountBalance'
-import { getAmmAccount } from '@play-money/accounts/lib/getAmmAccount'
-import { getUserAccount } from '@play-money/accounts/lib/getUserAccount'
 import type { SchemaResponse } from '@play-money/api-helpers'
 import { auth } from '@play-money/auth'
+import { getMarketBalance } from '@play-money/markets/lib/getMarketBalance'
+import { getUserBalanceInMarket } from '@play-money/markets/lib/getUserBalanceInMarket'
 import schema from './schema'
 
 export const dynamic = 'force-dynamic'
@@ -14,44 +13,17 @@ export async function GET(
 ): Promise<SchemaResponse<typeof schema.get.responses>> {
   try {
     const { id } = schema.get.parameters.parse(params)
-
-    const ammAccount = await getAmmAccount({ marketId: id })
-    const y = await getAccountBalance({
-      accountId: ammAccount.id,
-      currencyCode: 'YES',
-      marketId: id,
-      excludeTransactionTypes: ['MARKET_RESOLVE_LOSS', 'MARKET_RESOLVE_WIN'],
-    })
-    const n = await getAccountBalance({
-      accountId: ammAccount.id,
-      currencyCode: 'NO',
-      marketId: id,
-      excludeTransactionTypes: ['MARKET_RESOLVE_LOSS', 'MARKET_RESOLVE_WIN'],
-    })
-
-    let holdings = {}
     const session = await auth()
 
-    if (session?.user?.id) {
-      const userAccount = await getUserAccount({ id: session.user.id })
-      const yes = await getAccountBalance({ accountId: userAccount.id, currencyCode: 'YES', marketId: id })
-      const no = await getAccountBalance({ accountId: userAccount.id, currencyCode: 'NO', marketId: id })
+    const [balance, holdings] = await Promise.all([
+      getMarketBalance({
+        marketId: id,
+        excludeTransactionTypes: ['MARKET_RESOLVE_LOSS', 'MARKET_RESOLVE_WIN'],
+      }),
+      session?.user?.id ? getUserBalanceInMarket({ userId: session.user.id, marketId: id }) : {},
+    ])
 
-      holdings = {
-        YES: yes.toNumber(),
-        NO: no.toNumber(),
-      }
-    }
-
-    return NextResponse.json({
-      YES: y.toNumber(),
-      NO: n.toNumber(),
-      probability: {
-        YES: n.div(y.add(n)).toNumber(),
-        NO: y.div(y.add(n)).toNumber(),
-      },
-      holdings,
-    })
+    return NextResponse.json({ ...balance, holdings })
   } catch (error) {
     console.log(error) // eslint-disable-line no-console -- Log error for debugging
     return NextResponse.json({ error: 'Error processing request' }, { status: 500 })
